Don't error when trying to extend non-extensible errors
Before this change, `pRetry` blindly tries to modify the error thrown by the inner function. If that error happens to not be an extensible object, this modification itself throws, which obscures the inner error. This change makes `pRetry` not add these properties if the error object can't be extended under the presumption that retrying this error is better than hard erroring the first time its thrown.

Fixes #25